PIPS-S: Avoid solver recopy if bounds changed.
If current solver is dual simplex and bounds change, the solver
object in PIPSSInterface should not be recopied because changes
in bounds preserve dual feasibility.

To implement this change, add field to PIPSSInterface storing
solver type and add code to keep this field's state
consistent with the solver object in PIPSSInterface.