Fix error due to invoking size on lb, ub.
Fix errors in lines 834 and 1083 of PIPS/PIPS-S/Basic/BAData.cpp,
where lb and ub are mistakenly used instead of
l.getFirstStageVec() and u.getFirstStageVec(). In both places,
the assert is checking to make sure that the lengths of the lower
and upper bound vectors of first-stage variables are the same,
so that size changes are consistent.

Reported-by: Nai-Yuan Chiang <nychiang@mcs.anl.gov>