Cbc: Add patch for Cbc 2.9.8
Add patch for Cbc 2.7.5 for compatibility with PIPS-S. This patch adds
64-bit indexing to CoinIndexedArrays and fixes an issue with
infeasibilities in ClpSimplex::initialSolve; thus, the patches are really
to the CoinUtils and Cbc libraries.

The patch for Cbc 2.7.5 by Miles Lubin was then ported to Cbc 2.9.8 by
Geoff Oxberry, who committed the patch and assigned Miles Lubin as
author. The modifications required to port the patch were minor.

Signed-off-by: Geoffrey Oxberry <oxberry1@llnl.gov>